pyproject.toml
[project]
name = "jobrecommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.2.1",
    "groq>=0.32.0",
    "mcp[cli]>=1.18.0",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
